general: 3.8 reached EOL, remove legacy code and make pytz optional dependency
also enable 3.13 on CI
