+++ tox.ini
@@ -44,7 +44,7 @@ commands =
 deps =
     -e .[testing,optional]
 commands =
-    {envpython} -m mypy --install-types --non-interactive \
+    {envpython} -m mypy \
         -p {[testenv]package_name}       \
         # txt report is a bit more convenient to view on CI
         --txt-report  .coverage.mypy     \
